observability: add example server and client modes
