+++ observability/example/Makefile.env
@@ -1,6 +1,8 @@
 
 # Main Options
 
+#- Example mode: server or client (server,client) [server]
+MODE                 ?= server
 #- Generate and print config definition in given format and exit (default: '', means skip) (,json,md,mk) []
 CONFIG_GEN           ?=
 #- Dump config dest filename (string) []
@@ -25,6 +27,13 @@ HTTP_ETAG            ?= false
 HTTP_TLS_CERT        ?=
 #- KeyFile for serving HTTPS instead HTTP (string) []
 HTTP_TLS_KEY         ?=
+
+# Client Options
+
+#- HTTP server URL (string) [http://localhost:8080/demo]
+CLIENT_URL           ?= http://localhost:8080/demo
+#- HTTP request timeout (time.Duration) [3s]
+CLIENT_TIMEOUT       ?= 3s
 
 # OpenTelemetry Options
